Moved RulesDrawer to separate directory.
